% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjdotchart.R
\name{rjdotchart}
\alias{rjdotchart}
\title{This is a function which produces dot charts with right-justified y-axis labels}
\usage{
rjdotchart(x, labels = NULL, groups = NULL, gdata = NULL,
  cex = par("cex"), pt.cex = cex, pch = 21, gpch = 21, bg = par("bg"),
  color = par("fg"), gcolor = par("fg"), lcolor = "gray",
  xlim = range(x[is.finite(x)]), main = NULL, xlab = NULL, ylab = NULL,
  ...)
}
\arguments{
\item{x}{either a vector or a matrix of numeric calues (NAs are allowed).}

\item{labels}{a vector of labels for each point. For vectors the default is to use names(x) and for matrices the row labels dimnames(x)[[1]].}

\item{groups}{an optional factor indicating how the elements of x are grouped. If x is a matrix, groups will default to the columns of x.}

\item{gdata}{data values for the groups. This is typically a summary such as the median or mean of each group.}

\item{cex}{the character size to be used. Setting cex to a value smaller than one can be a useful way of avoiding label overlap. Unlike many other graphics functions, this sets the actual size, not a multiple of par("cex").}

\item{pt.cex}{the cex to be applied to plotting symbols. This behaves like cex in plot().}

\item{pch}{the plotting character or symbol to be used}

\item{gpch}{the plotting character or symbol to be used for group values}

\item{bg}{the background color of plotting characters or symbols to be used}

\item{color}{the color(s) to be used for group labels and values}

\item{gcolor}{the single color to be used for group labels and values}

\item{lcolor}{the colors to be used for the horizontal lines}

\item{xlim}{horizontal range for the plot}

\item{main}{overall title for the plot}

\item{xlab}{axis notations as in title}

\item{xlab}{axis notations as in title}
}
\value{
This function is invoked for its side effect, which is to produce two variatns of dotplots, with right-justified y-axis labels
}
\description{
This is a function which produces dot charts with right-justified y-axis labels
}
\examples{
rjdotchart(VADeaths[1,], main = "Death Rates in Virginia - 1940")
}
